refactor(mis-web,portal-web): optimize structure
