+++ apps/mis-web/src/layouts/base/header/index.tsx
@@ -3,9 +3,9 @@ import { Space, Typography } from "antd";
 import Link from "next/link";
 import { join } from "path";
 import React from "react";
-import { DefaultClusterSelector } from "src/components/ClusterSelector";
 import { antdBreakpoints } from "src/layouts/base/constants";
 import { BigScreenMenu } from "src/layouts/base/header/BigScreenMenu";
+import { DefaultClusterSelector } from "src/layouts/base/header/DefaultClusterSelector";
 import { Logo } from "src/layouts/base/header/Logo";
 import { NavItemProps } from "src/layouts/base/NavItemProps";
 import { User } from "src/stores/UserStore";
